+++ src/app/constants/constants.ts
@@ -1,3 +1,5 @@
+import { type } from "os";
+
 /* eslint-disable @typescript-eslint/naming-convention */
 export const AVALIABLE_LANGUAJES = ['en', 'es'];
 export const DATABASE_NAME = 'AureoDatabase.db';
@@ -67,3 +69,20 @@ export const LIST_TYPES = {
 export const FILES = {
   GIFTS: '../assets/i18n/gifts.es.json'
 }
+
+export enum GiftTypesEnum  {
+  COMMON_GIFTS =  'COMMON_GIFTS',
+  CURSES =  'CURSES',
+  DIVINE_GIFTS = 'DIVINE_GIFTS'
+}
+
+export type GiftType = GiftTypesEnum.COMMON_GIFTS | GiftTypesEnum.CURSES | GiftTypesEnum.DIVINE_GIFTS;
+
+export enum GiftSubTypesEnum {
+  PHYSICAL =  "PHYSICAL",
+  MENTAL = "MENTAL",
+  SUPERNATURAL = "SUPERNATURAL",
+  SOCIAL = "SOCIAL"
+}
+
+export type GiftSubType = GiftSubTypesEnum.PHYSICAL | GiftSubTypesEnum.MENTAL | GiftSubTypesEnum.SUPERNATURAL | GiftSubTypesEnum.SOCIAL;